Pin `@radix-ui/react-slot` dependency to SemVer MAJOR
Pinning it to its patch version contributes to dependency duplication in applications using `basehub` and `@radix-ui/react-slot`.
The package is not known for violating SemVer so pinning the version to its major is safe.

This ensures fixes are propagated faster throughout the ecosystem. In this case, `@radix-ui/slot` published a fix for React 19 compat that `basehub` users won't get until a new `basehub` release is out. It helps us in the React team in gathering feedback quicker since we don't need to ensure each dependent updates its dependencies even when their dependencies can ensure compatibility with a SemVer MINOR.